nulls in catalog response
node server.js --port 8888 -f metadata/TestData2.0.json -f metadata/SSCWeb.json
curl http://localhost:8888/SSCWeb/hapi/catalog

gave a response with [{id: "ace", title: "ACE"}, null,{id: "aim", title: "AIM"}]

The cause of this was a variable name that was not updated in checkinfo():

if (i < xcatalog.length - 1) {

The variable i was not defined and it is surprising that the code almost worked anyway.

This commit changes the variable name associated with the array element in the /catalog response.